--- parallel-consumer-core/src/main/java/io/confluent/csid/actors/Interruptible.java
@@ -1,28 +1,0 @@
-package io.confluent.csid.actors;
-
-import lombok.AllArgsConstructor;
-import lombok.EqualsAndHashCode;
-import lombok.RequiredArgsConstructor;
-import lombok.ToString;
-
-/**
- * todo docs
- *
- * @author Antony Stubbs
- */
-public interface Interruptible {
-    void interruptProcessBlockingMaybe(Reason reason);
-
-    /**
-     * todo docs
-     */
-    @AllArgsConstructor
-    @RequiredArgsConstructor
-    @ToString
-    @EqualsAndHashCode
-    class Reason {
-        public static final Reason UNKNOWN = new Reason("unknown-reason");
-        private final String desc;
-        private Reason root;
-    }
-}
