+++ parallel-consumer-core/src/main/java/io/confluent/parallelconsumer/internal/AbstractParallelEoSStreamProcessor.java
@@ -5,7 +5,7 @@ package io.confluent.parallelconsumer.internal;
  */
 
 import io.confluent.csid.actors.Actor;
-import io.confluent.csid.actors.Interruptible.Reason;
+import io.confluent.csid.actors.Interruptable.Reason;
 import io.confluent.csid.utils.TimeUtils;
 import io.confluent.parallelconsumer.ErrorInUserFunctionException;
 import io.confluent.parallelconsumer.ParallelConsumer;
@@ -1120,7 +1120,7 @@ public abstract class AbstractParallelEoSStreamProcessor<K, V> implements Parall
      */
     public void notifySomethingToDo(Reason reason) {
         // todo reason enum? extend? e.g. Reason.COMMIT_TIME ?
-        getMyActor().interruptProcessBlockingMaybe(reason);
+        getMyActor().interruptProcessAsync(reason);
     }
 
     @Override
